+++ VDA-DID-Registry/ChangeLog.md
@@ -1,3 +1,7 @@
+2023-04-03 (v0.2.1)
+-------------------
+- Added `isRegistered()` function for `VDAXPReward` contract
+
 2023-04-03 (v0.2.0)
 -------------------
 Updates on audit report